Warning fixes:         * class.c (make_method_vec): Cast 1st argument of `bzero' to (PTR).         (add_method): Likewise for arguments 1 & 2 of `bcopy'.         * decl.c (signal_catch): Mark with ATTRIBUTE_NORETURN.         * pt.c (process_partial_specialization): Cast 1st argument of         `bzero' to (PTR).         * tree.c (build_base_fields): Cast `base_align' to (int) when         comparing against one.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@23156 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cp/tree.c
@@ -806,7 +806,7 @@ build_base_fields (rec)
 	  base_align = MAX (base_align, DECL_ALIGN (decl));
 	  DECL_SIZE (decl)
 	    = size_int (MAX (TREE_INT_CST_LOW (DECL_SIZE (decl)),
-			     base_align));
+			     (int) base_align));
 	}
       else if (DECL_SIZE (decl) == integer_zero_node)
 	saw_empty = 1;
